chore: No more "module-inception"
Previously, as an artifact (I think) left-over from when we merged all the
previously-distinct crates comprising Hipcheck into one crate, we had a
"module-inception" problem that Clippy complained about. Besides Clippy
complaining, it also just made for some awkward imports.

This commit removes this problem, moving the contents of the doubly-named
modules up to their parent module, and updating imports to use the new
locations.
