+++ hipcheck/src/analysis/report_builder.rs
@@ -8,8 +8,8 @@ use crate::error::Result;
 use crate::hc_error;
 use crate::report::Concern;
 pub use crate::report::*;
-use crate::session::session::Session;
-use crate::source::source::SourceQuery;
+use crate::session::Session;
+use crate::source::SourceQuery;
 use crate::version::VersionQuery;
 use std::default::Default;
 use std::result::Result as StdResult;
