+++ hipcheck/src/cache/mod.rs
@@ -1,4 +1,3 @@
-#[allow(clippy::module_inception)]
 #[allow(dead_code)]
 pub mod plugin_cache;
 pub mod repo_cache;
